home/steam_bot/requirements.txt
PySocks
python-telegram-bot==12.0.0b1
requests
